requirements.txt
dash==2.14.2
dash-bootstrap-components==1.5.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
gunicorn==21.2.0
plotly==5.18.0
opencv-contrib-python==4.8.1.78
ultralytics==8.0.196
numpy>=1.26.0
torch==2.2.0
torchvision==0.17.0
scipy==1.11.3 